Extract log level from the stdout
